+++ app/src/main/java/com/example/groceryguruapp/CreateAccount.kt
@@ -1,0 +1,33 @@
+package com.example.groceryguruapp
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+
+
+
+class CreateAccount: Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.create_account, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        // view.findViewById<Button>(R.id.btn_login).setOnClickListener {
+        ///goes to home screen
+        ///().navigate(R.id.action_FirstFragment_to_SecondFragment)
+
+        // }
+
+
+    }
+}
