+++ app/src/main/java/com/example/groceryguruapp/SecondFragment.kt
@@ -25,7 +25,7 @@ class SecondFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         view.findViewById<Button>(R.id.button_second).setOnClickListener {
-            findNavController().navigate(R.id.action_SecondFragment_to_FirstFragment)
+          //  findNavController().navigate(R.id.action_SecondFragment_to_FirstFragment)
         }
     }
 }
