--- app/src/main/java/com/example/groceryguruapp/createAccount.kt
@@ -1,31 +1,0 @@
-package com.example.groceryguruapp
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Button
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
-
-class createAccount : Fragment(){
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.login, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        view.findViewById<Button>(R.id.btn_login).setOnClickListener {///goes to home screen
-            ///().navigate(R.id.action_FirstFragment_to_SecondFragment)
-
-            view.findViewById<Button>(R.id.link_signup).setOnClickListener {/////goes to profile
-              //  findNavController().navigate(R.id.action_FirstFragment_to_SecondFragment)
-            }
-        }
-    }
-}
