+++ app/src/main/java/com/example/groceryguruapp/login.kt
@@ -22,13 +22,15 @@ import androidx.navigation.fragment.findNavController
        override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
            super.onViewCreated(view, savedInstanceState)
 
-           view.findViewById<Button>(R.id.btn_login).setOnClickListener {///goes to home screen
+          // view.findViewById<Button>(R.id.btn_login).setOnClickListener {
+               ///goes to home screen
                ///().navigate(R.id.action_FirstFragment_to_SecondFragment)
 
+          // }
                view.findViewById<Button>(R.id.link_signup).setOnClickListener {/////goes to profile
                    findNavController().navigate(R.id.action_login_to_createAccount)
                }
-           }
+
        }
    }
