+++ app/src/main/res/layout/login.xml
@@ -8,7 +8,7 @@
         android:orientation="vertical"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:paddingTop="275dp"
+        android:paddingTop="240dp"
         android:paddingLeft="24dp"
         android:paddingRight="24dp">
 
@@ -50,10 +50,10 @@
             android:padding="12dp"
             android:text="Login"/>
 
-        <TextView
+        <Button
             android:id="@+id/link_signup"
             android:layout_width="fill_parent"
-            android:layout_height="11dp"
+            android:layout_height="wrap_content"
             android:layout_marginBottom="24dp"
             android:gravity="center"
             android:text="No account yet? Create one"
